% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory}
\usage{
create_dir(key, ...)
}
\arguments{
\item{key}{(character) A key name. Optional.}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}
}
\description{
Create a directory
}
\examples{
\dontrun{
# Make a key
create_dir("newdir")
}
}

