% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etseed-package.R
\docType{package}
\name{etseed-package}
\alias{etseed}
\alias{etseed-package}
\title{etcd R client}
\description{
etcd client for R
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

