% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/keys.R
\name{keys}
\alias{key}
\alias{keys}
\title{List a key or all keys}
\usage{
keys(...)

key(key, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[httr]{GET}}}

\item{key}{(character) A key name. Optional.}
}
\description{
List a key or all keys
}
\examples{
\dontrun{
# List all keys
keys()

# A single key
key("mykey")
}
}

