% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members}
\alias{member_add}
\alias{member_change}
\alias{member_delete}
\alias{member_list}
\alias{members}
\title{Manage etcd members}
\usage{
member_list(...)

member_add(id, ...)

member_change(id, newid, ...)

member_delete(id, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[httr]{GET}}}

\item{id}{(character) A member id}

\item{newid}{(logical) new member id}
}
\value{
Logical or R list
}
\description{
Manage etcd members
}
\examples{
\dontrun{
Sys.setenv(ETSEED_USER = "root")
Sys.setenv(ETSEED_PWD = "pickbetterpwd")

# list members
member_list()

# add a member
member_add("http://10.0.0.10:2380")

# change a member
mms <- member_list()
member_change(mms$members[[1]]$id, "http://10.0.0.10:8380", config=verbose())

# delete a member
mms <- member_list()
member_delete(mms$members[[1]]$id)
}
}

