% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Get etcd metrics}
\arguments{
\item{pretty}{(logical) Print easier to read with newlines, or as a single string.
Default: \code{TRUE}}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}
}
\value{
Prints a human readable text representation to console
of the metrics of your etcd cluster
}
\description{
Get etcd metrics
}
\examples{
\dontrun{
# make a client
cli <- etcd()

cli$metrics()
cli$metrics(pretty = FALSE)
}
}

