% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Get etcd statistics}
\usage{
stats(which = "leader", pretty = TRUE, ...)
}
\arguments{
\item{which}{(character) one of leader (default), self, or store}

\item{pretty}{(logical) parse text to an R list, or not. Default: \code{TRUE}}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}
}
\value{
Prints a human readable text representation to console.
}
\description{
Get etcd statistics
}
\examples{
\dontrun{
# leader stats
stats()

# self stats
stats("self")

# store stats
stats("store")

# Prety or not
stats(pretty = TRUE)
stats(pretty = FALSE)
}
}

