% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Get etcd version}
\usage{
version(...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[httr]{GET}}}
}
\description{
Get etcd version
}
\examples{
\dontrun{
version()
}
}

