% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\alias{auth_disable}
\alias{auth_enable}
\alias{auth_status}
\title{etcd authentication - auth control}
\usage{
auth_status(...)

auth_enable(...)

auth_disable(auth_user, auth_pwd, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[httr]{GET}},
\code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}

\item{auth_user,auth_pwd}{(character) Username and password for the authenticated user,
the root user}
}
\description{
etcd authentication - auth control
}
\examples{
\dontrun{
# check authentication status
auth_status()

# enable authentication
auth_enable()

# woops, if you got error about root user, craeate one first
add_user(user = "root", password = "pickbetterpwd")

# disable authentication
auth_disable("root", "pickbetterpwd")

# check again, now disabled
auth_status()

# Users
# Add user
user_add("jane", "janepwd", "root", "pickbetterpwd")
# List users
user_list()
# Get a single user
user_get("root")
user_get("jane")
# Delete user
user_delete("jane", "root", "pickbetterpwd")
}
}

