% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/etseed-package.R
\docType{package}
\name{etseed-package}
\alias{etseed}
\alias{etseed-package}
\title{etcd R client}
\description{
Ping urls to time requests
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

