% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etseed-package.R
\docType{package}
\name{etseed-package}
\alias{etseed-package}
\alias{etseed}
\title{etcd client for R}
\description{
etcd client for R
}
\section{etcd keys}{

Note that all keys have to be prefixed by a forward slash (\code{/})
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
