% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/keys.R
\name{keys}
\alias{create}
\alias{create_inorder}
\alias{delete}
\alias{key}
\alias{keys}
\alias{update}
\title{List a key or all keys}
\usage{
keys(recursive = NULL, sorted = NULL, ...)

key(key, recursive = NULL, sorted = NULL, ...)

create(key, value, ttl = NULL, ...)

update(key, value, ttl = NULL, ...)

create_inorder(key, value, ttl = NULL, ...)

delete(key, prevValue = NULL, prevIndex = NULL, ...)
}
\arguments{
\item{recursive}{Whether to do recursive something???}

\item{sorted}{Whether to return sorted or not.}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}

\item{key}{(character) A key name. Optional.}

\item{value}{Any object to store. Required for \code{\link{create}}, \code{\link{update}},
and \code{\link{delete}} functions}

\item{ttl}{(integer) Seconds after which the key will be removed.}

\item{prevValue}{(character) Previous value to match against.}

\item{prevIndex}{(integer) Previous index to match against.}
}
\description{
List a key or all keys
}
\details{
\code{\link{create}} and \code{\link{update}} are essentially the same thing, but
get different names so as not to confuse people (eg. if create did create and update functions.)
}
\examples{
\dontrun{
# Make a key
create(key="mykey", value="this is awesome")
create(key="things", value="and stuff!")
## use ttl
create(key="stuff", value="tables", ttl=10)

# Update a key
update(key="things", value="and stuff! and more things")

# Create an in-order key
create_inorder("queue", "thing1")
create_inorder("queue", "thing2")
create_inorder("queue", "thing3")
key("queue", sorted = TRUE, recursive = TRUE)

# List all keys
keys()
keys(sorted = TRUE)
keys(recursive = TRUE)
keys(sorted = TRUE, recursive = TRUE)

# List a single key
key("mykey")
key("things")

# Delete a key
create("hello", "world")
delete("hello")
## Delete only if matches previous value, fails
delete("things", prevValue="two")
## Delete only if matches previous index
### Fails
delete("things", prevIndex=1)
### Works
delete("things", prevIndex=13)

# curl options
library("httr")
keys(config = verbose())
}
}

