% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members}
\alias{members}
\title{Manage etcd members}
\arguments{
\item{id}{(character) A member id}

\item{newid}{(logical) new member id}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}
}
\value{
Logical or R list
}
\description{
Manage etcd members
}
\examples{
\dontrun{
Sys.setenv(ETSEED_USER = "root")
Sys.setenv(ETSEED_PWD = "pickbetterpwd")

# make a client
cli <- etcd()

# list members
cli$member_list()

# add a member
cli$member_add("http://10.0.0.10:2380")

# change a member
mms <- cli$member_list()
cli$member_change(mms$members[[1]]$id, "http://10.0.0.10:8380", config=verbose())

# delete a member
mms <- cli$member_list()
cli$member_delete(mms$members[[1]]$id)
}
}

