% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Get etcd metrics}
\usage{
metrics(pretty = TRUE, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[httr]{GET}}}
}
\value{
Prints a human readable text representation to console.
}
\description{
Get etcd metrics
}
\examples{
\dontrun{
metrics()
metrics(FALSE)
}
}

