% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{roles}
\alias{role_add}
\alias{role_delete}
\alias{role_get}
\alias{role_list}
\alias{roles}
\title{etcd authentication - roles}
\usage{
role_add(role, perm_read = NULL, perm_write = NULL, grant_read = NULL,
  grant_write = NULL, revoke_read = NULL, revoke_write = NULL, auth_user,
  auth_pwd, ...)

role_list(...)

role_get(role, ...)

role_delete(role, auth_role, auth_pwd, ...)
}
\arguments{
\item{role}{(character) Role to act upon}

\item{perm_read}{Read permission to set}

\item{perm_write}{Write permission to set}

\item{grant_read}{Grant read permission to set}

\item{grant_write}{Grant write permission to set}

\item{revoke_read}{Revoke read permission to set}

\item{revoke_write}{Revoke write permission to set}

\item{auth_user, auth_pwd}{(character) Username and password for the authenticated user,
the root user}

\item{...}{Further args passed on to \code{\link[httr]{GET}},
\code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}
}
\description{
etcd authentication - roles
}
\examples{
\dontrun{
# Add role
perms <- list(fun = list(kv = list(read = "/message/*")))
role_add("fun", perms, "root", "pickbetterpwd")

# List roles
role_list()

# Get a single role
role_get("root")
role_get("fun")

# Delete role
role_delete("fun", "root", "pickbetterpwd")

# Update a role
### FIXME - still working on this
## First, create
role_add("stuff", perm_read = "/message/*", perm_write = "/message/*",
   auth_user = "root", auth_pwd = "pickbetterpwd")
## udpate
"xxx"
## get
role_get("stuff")
}
}

