% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{user_add}
\alias{user_delete}
\alias{user_get}
\alias{user_list}
\alias{users}
\title{etcd authentication - users}
\usage{
user_add(user, password, auth_user = NULL, auth_pwd = NULL, roles = NULL,
  ...)

user_list(...)

user_get(user, ...)

user_delete(user, auth_user, auth_pwd, ...)
}
\arguments{
\item{user}{(character) User name to create/delete}

\item{password}{(character) Password to give the new user}

\item{auth_user, auth_pwd}{(character) Username and password for the authenticated user,
the root user}

\item{roles}{(list/character vector) Roles to give the new user}

\item{...}{Further args passed on to \code{\link[httr]{GET}},
\code{\link[httr]{PUT}}, or \code{\link[httr]{DELETE}}}
}
\description{
etcd authentication - users
}
\examples{
\dontrun{
# Add user
user_add("jane", "janepwd", "root", "pickbetterpwd")

# List users
user_list()

# Get a single user
user_get("root")
user_get("jane")

# Delete user
user_delete("jane", "root", "pickbetterpwd")
}
}

